http://code.google.com/p/mybatis/issues/detail?id=573 . fix deadlock and added a test

+++ src/test/java/org/apache/ibatis/submitted/cache/CreateDB.sql
@@ -14,7 +14,7 @@
 --    limitations under the License.
 --
 
-drop table person;
+drop table person if exists;
 
 create table person(
 	id int,
